Rename actions/ -> endpointActions/
