+++ CMakeLists.txt
@@ -120,18 +120,18 @@ set(FILES
         bridge/StreamDescription.h
         bridge/TransportDescription.h
         bridge/VideoStream.h
-        bridge/actions/ActionContext.h
-        bridge/actions/ApiActions.h
-        bridge/actions/GetConferenceInfo.cpp
-        bridge/actions/GetEndpointInfo.cpp
-        bridge/actions/AllocateConference.cpp
-        bridge/actions/ConferenceActions.cpp
-        bridge/actions/GetConferences.cpp
-        bridge/actions/About.cpp
-        bridge/actions/GetStats.cpp
-        bridge/actions/ApiHelpers.cpp
-        bridge/actions/ApiHelpers.h
-        bridge/actions/BarbellActions.cpp
+        bridge/endpointActions/ActionContext.h
+        bridge/endpointActions/ApiActions.h
+        bridge/endpointActions/GetConferenceInfo.cpp
+        bridge/endpointActions/GetEndpointInfo.cpp
+        bridge/endpointActions/AllocateConference.cpp
+        bridge/endpointActions/ConferenceActions.cpp
+        bridge/endpointActions/GetConferences.cpp
+        bridge/endpointActions/About.cpp
+        bridge/endpointActions/GetStats.cpp
+        bridge/endpointActions/ApiHelpers.cpp
+        bridge/endpointActions/ApiHelpers.h
+        bridge/endpointActions/BarbellActions.cpp
         bridge/engine/ActiveMediaList.cpp
         bridge/engine/ActiveMediaList.h
         bridge/engine/AudioForwarderReceiveJob.cpp
