+++ bridge/endpointActions/AllocateConference.cpp
