+++ bridge/endpointActions/GetConferenceInfo.cpp
