+++ bridge/endpointActions/GetConferences.cpp
