+++ bridge/endpointActions/GetEndpointInfo.cpp
